Convert Stat Components to Struct (#557)
convert stat components to follow the struct pattern outlined in
`contracts/src/components/types/StatComponent.sol`. this includes:
- HealthComponent (uses `Stat.sync` field rather than HealthCurrent)
- PowerComponent
- ViolenceComponent
- HarmonyComponent
- SlotsComponent
- StaminaComponent (uses `Stat.sync` field rather than StaminaCurrent)

notes in that file elaborate on how this struct is used across both Traits and 
various items. sc-logic, tests and FE have been updated to match this new
pattern

+++ packages/contracts/src/systems/ProductionLiquidateSystem.sol
@@ -47,10 +47,7 @@ contract ProductionLiquidateSystem is System {
     uint256 nodeID = LibProduction.getNode(components, productionID);
     uint256 targetNodeID = LibProduction.getNode(components, targetProductionID);
     require(nodeID == targetNodeID, "FarmLiquidate: target too far");
-    require(
-      LibAccount.sharesRoom(components, accountID, nodeID),
-      "FarmLiquidate: node too far"
-    );
+    require(LibAccount.sharesRoom(components, accountID, nodeID), "FarmLiquidate: node too far");
 
     // check that the pet is capable of liquidating the target production
     uint256 targetPetID = LibProduction.getPet(components, targetProductionID);
@@ -65,7 +62,7 @@ contract ProductionLiquidateSystem is System {
     uint256 bounty = LibPet.calcBounty(components, petID, balance);
     uint256 recoil = LibPet.calcDrain(components, petID, bounty);
     LibCoin.inc(components, productionID, bounty);
-    LibPet.drain(components, petID, recoil);
+    LibPet.drain(components, petID, int32(int(recoil)));
 
     // kill the target and shut off the production
     LibPet.kill(components, targetPetID);
